add statesync protocol version
