+++ libs/statesync/src/monad/statesync/statesync_client_context.cpp
@@ -1,16 +1,20 @@
 #include <monad/core/block.hpp>
 #include <monad/db/util.hpp>
 #include <monad/mpt/ondisk_db_config.hpp>
+#include <monad/mpt/update.hpp>
+#include <monad/statesync/statesync_client.h>
 #include <monad/statesync/statesync_client_context.hpp>
+#include <monad/statesync/statesync_protocol.hpp>
 
+#include <deque>
 #include <sys/sysinfo.h>
 
 using namespace monad;
+using namespace monad::mpt;
 
 monad_statesync_client_context::monad_statesync_client_context(
     std::vector<std::filesystem::path> const dbname_paths,
-    std::filesystem::path const genesis, uint8_t const prefix_bytes,
-    monad_statesync_client *const sync,
+    std::filesystem::path const genesis, monad_statesync_client *const sync,
     void (*statesync_send_request)(
         struct monad_statesync_client *, struct monad_sync_request))
     : db{machine,
@@ -24,9 +28,9 @@ monad_statesync_client_context::monad_statesync_client_context(
              .dbname_paths = dbname_paths}}
     , tdb{db}
     , progress(
-          static_cast<size_t>(std::pow(16, prefix_bytes * 2)),
+          MONAD_STATESYNC_N_PREFIXES,
           {db.get_latest_block_id(), db.get_latest_block_id()})
-    , prefix_bytes(prefix_bytes)
+    , protocol(MONAD_STATESYNC_N_PREFIXES)
     , target{db.get_latest_block_id()}
     , current{db.get_latest_block_id() == mpt::INVALID_BLOCK_ID ? 0 : db.get_latest_block_id() + 1}
     , expected_root{NULL_ROOT}
@@ -35,4 +39,79 @@ monad_statesync_client_context::monad_statesync_client_context(
     , sync{sync}
     , statesync_send_request{statesync_send_request}
 {
+}
+
+void monad_statesync_client_context::commit()
+{
+    std::deque<mpt::Update> alloc;
+    std::deque<byte_string> bytes_alloc;
+    std::deque<hash256> hash_alloc;
+    UpdateList accounts;
+    for (auto const &[addr, delta] : deltas) {
+        UpdateList storage;
+        std::optional<byte_string_view> value;
+        if (delta.has_value()) {
+            auto const &[acct, deltas] = delta.value();
+            value = bytes_alloc.emplace_back(encode_account_db(addr, acct));
+            for (auto const &[key, val] : deltas) {
+                storage.push_front(alloc.emplace_back(Update{
+                    .key = hash_alloc.emplace_back(keccak256(key.bytes)),
+                    .value = val == bytes32_t{}
+                                 ? std::nullopt
+                                 : std::make_optional<byte_string_view>(
+                                       bytes_alloc.emplace_back(
+                                           encode_storage_db(key, val))),
+                    .incarnation = false,
+                    .next = UpdateList{},
+                    .version = static_cast<int64_t>(current)}));
+            }
+        }
+        accounts.push_front(alloc.emplace_back(Update{
+            .key = hash_alloc.emplace_back(keccak256(addr.bytes)),
+            .value = value,
+            .incarnation = false,
+            .next = std::move(storage),
+            .version = static_cast<int64_t>(current)}));
+    }
+    UpdateList code_updates;
+
+    ankerl::unordered_dense::segmented_set<bytes32_t> remaining;
+    std::deque<bytes32_t> upserted;
+    for (auto const &hash : hash) {
+        if (code.contains(hash)) {
+            code_updates.push_front(alloc.emplace_back(Update{
+                .key = NibblesView{hash},
+                .value = code.at(hash),
+                .incarnation = false,
+                .next = UpdateList{},
+                .version = static_cast<int64_t>(current)}));
+            upserted.emplace_back(hash);
+        }
+        else {
+            remaining.insert(hash);
+        }
+    }
+
+    auto state_update = Update{
+        .key = state_nibbles,
+        .value = byte_string_view{},
+        .incarnation = false,
+        .next = std::move(accounts),
+        .version = static_cast<int64_t>(current)};
+    auto code_update = Update{
+        .key = code_nibbles,
+        .value = byte_string_view{},
+        .incarnation = false,
+        .next = std::move(code_updates),
+        .version = static_cast<int64_t>(current)};
+    UpdateList updates;
+    updates.push_front(state_update);
+    updates.push_front(code_update);
+    db.upsert(std::move(updates), current, false, false);
+    tdb.set_block_number(current);
+    for (auto const &hash : upserted) {
+        MONAD_ASSERT(code.erase(hash) == 1);
+    }
+    hash = std::move(remaining);
+    deltas.clear();
 }
