+++ libs/statesync/src/monad/statesync/statesync_version.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <cstdint>
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+// Bump when there are modifications to the protocol
+constexpr uint32_t MONAD_STATESYNC_VERSION = 1;
+
+#ifdef __cplusplus
+}
+#endif
